requirements.txt
argparse
numpy
opencv-python
pyyaml
ultralytics
onnxruntime
